Refactored Symptoms, Food and Additional Data to .tsx
