+++ src/AdditionalData.tsx
@@ -2,8 +2,22 @@ import React, { useState, useEffect } from "react";
 import { toast } from "react-toastify";
 import firebase from "./firebase.js";
 
-function AdditionalData(props) {
-  const [additionalData, setAdditionalData] = useState({
+type AdditionalDataProps = {
+  userID: string;
+  kidId: string;
+  date: Date;
+  date_yyyy_mm: string;
+};
+
+type TAdditionalData = {
+  outdoor:string,
+  notes:string,
+  ac: boolean,
+  nails: boolean,
+};
+
+export const  AdditionalData:React.FC<AdditionalDataProps> = (props) => {
+  const [additionalData, setAdditionalData] = useState<TAdditionalData>({
     outdoor: "",
     notes: "",
     ac: false,
@@ -45,7 +59,7 @@ function AdditionalData(props) {
     });
   }, [props]);
 
-  const handleChange = (e) => {
+  const handleChange = (e: React.ChangeEvent<any>) => {
     const { name, value, type, checked } = e.target;
 
     type === "checkbox"
@@ -53,7 +67,7 @@ function AdditionalData(props) {
       : setAdditionalData({ ...additionalData, [name]: value });
   };
 
-  const saveAdditionalData = (e, additionalData) => {
+  const saveAdditionalData = (e:React.FormEvent<HTMLFormElement>, additionalData:TAdditionalData) => {
     e.preventDefault();
 
     if (firebaseID === "") {
@@ -63,7 +77,7 @@ function AdditionalData(props) {
     } else {
       /* update data */
 
-      let updates = {};
+      let updates: any = {};
       updates["/" + firebaseID] = additionalData;
       additionalDataRef.update(updates);
     }
@@ -102,7 +116,7 @@ function AdditionalData(props) {
             <input
               type="text"
               name="outdoor"
-              maxLength="100"
+              maxLength={100}
               value={additionalData.outdoor}
               onChange={(e) => handleChange(e)}
             />
@@ -111,8 +125,8 @@ function AdditionalData(props) {
             <label htmlFor="notes">Additional notes</label>
             <textarea
               name="notes"
-              rows="4"
-              maxLength="200"
+              rows={4}
+              maxLength={200}
               className="notes"
               value={additionalData.notes}
               onChange={(e) => handleChange(e)}
@@ -135,4 +149,4 @@ function AdditionalData(props) {
   );
 }
 
-export default AdditionalData;
+
